Finished base game

+++ Assets/Example/Script/Scene/Gameplay/Module/Character/Model/ICharacterModel.cs
@@ -10,7 +10,7 @@ namespace Example.Scene.Gameplay.Character
         public float Energy {get ;}
         public float Bladder {get ;}
         public float Hunger { get;}
-        public int Money { get;}
+        public float Money { get;}
         public Vector3 Position {get;}
         public bool IsWalking {get;}
     }
